Add the Spring IO plugin and align with the Platform where possible
Configure the Spring IO plugin such that it's only applied when the
build is run with -PplatformVersion=<version>. This platformVersion
property is used to determine the version of the Platform that will
be used when running the springIoCheck task. The plugin can be used
by running a build as follows:

./gradlew clean springIoCheck -PplatformVersion=1.0.1.RELEASE -PJDK7_HOME=… -PJDK8_HOME=…

This will test the project on JDK7 and JDK 8 using the dependencies
defined in Spring IO Platform 1.0.1.RELEASE.

Running the plugin identified a dependency on CGLib (use of
spring-core's repackaged CGLib is preferred), but the dependency
appears to be unused. It has been removed.

The versions of a number of other dependencies have been updated to
use their latest versions and align them with Spring IO Platform. Most
notable is probably a move from Guava 15.0 to 17.0.