Allow empty input with double quotes.
Empty input with double quotes("") was not allowed because of filtering empty string in StandardParameterResolver.resolve().
Assuming ShellMethod test(String a, @ShellOption(defaultValue = "") String b) and input 'test "" B'. Expected parameter is (a="", b="B"). But result is (a="B", b="").
So I remove filtering in StandardParameterResolver.resolve().

And remove unnecessary code.