Update ShellRunner#run to return void
Before this commit, `ShellRunner#run` was returning a boolean (and not
void as it was before). The javadoc mentions that the "return value
indicates if run operation happened and no further runners should be
used", but it is not the ShellRunner's responsibility to decide the flow
of execution (ie if another runner should run or not). Custom implementations
have no context to decide to return true or false.