Redesign command programming model
This commit revisits the previous programming model in order to fix
several design issues and inconsistencies. The major changes are:

- Revisit declarative and programmatic command definition
- Redesign command registration and discovery
- Introduce new adapters to create commands from annotated methods
- Fix incorrect implementation of non-interactive shells #1218
- Remove Shell runner precedence #1219
- Make interactive mode the default #1186
- Remove CommandResolver which is redundant #1217
- Make it easier to use Spring Shell without Spring Boot #1207
- Revisit exceptions

I tried to introduce these changes in an incremental way, but APIs
were so entangled together that it was very hard to break them down
to smaller change sets.

This commit does not address Spring Boot support on purpose.
Support for Spring Boot will be addressed in a separate commit.

Resolves #1186
Resolves #1207
Resolves #1217
Resolves #1218
Resolves #1219