Replace Stacktrace command with debug mode
The stacktrace command requires global mutable
state to keep the stacktrace of the last command.
There is no good place for this global state and
accessing it from the shell is not convenient.

Similar to many popular shell frameworks, this commit
introduce a debug mode, which when enabled, prints the
stack trace of the last command right away instead of
the user asking for it right after.